Fix typo in call to Parse.User.become(response.sessionToken)
The existing code tries to become the user for the session token `response.sessionToken` but this will always be `undefined` since `sessionToken` isn't a property of `ParseUser`. Calling `response.getSessionToken()` gets the correct session token.

The reason this didn't manifest before is that the code later in the middleware works regardless of whether we become the user.